Move minimize_bfgs.py to _bfgs.py
