+++ jax/scipy/optimize/_bfgs.py
